+++ test/directive_test.dart
@@ -83,6 +83,14 @@ void main() {
   });
 
   group('problem cases:', () {
+    test('Deprecated unquoted default region name', () {
+      final d = new Directive.tryParse("#docregion ,a");
+      expect(d.kind, Kind.startRegion);
+      expect(d.rawArgs, ",a");
+      expect(d.args, ['', 'a']);
+      expect(d.issues, ['unquoted default region name is deprecated']);
+    });
+
     test('Duplicate "a" region', () {
       final d = new Directive.tryParse('#docregion a,b,c,a');
       expect(d.kind, Kind.startRegion);
@@ -92,9 +100,9 @@ void main() {
     });
 
     test('Duplicate "" region', () {
-      final d = new Directive.tryParse('#docregion ,');
+      final d = new Directive.tryParse("#docregion '',''");
       expect(d.kind, Kind.startRegion);
-      expect(d.rawArgs, ',');
+      expect(d.rawArgs, "'',''");
       expect(d.args, ['']);
       expect(d.issues, ['repeated argument ""']);
     });